Associate feedback_files by optional test_run_id

+++ app/controllers/api/feedback_files_controller.rb
@@ -3,17 +3,23 @@ module Api
   # Allows for pushing and downloading of Feedback Files
   # Uses Rails' RESTful routes (check 'rake routes' for the configured routes)
   class FeedbackFilesController < MainApiController
+    include DownloadHelper
+
     # Define default fields for index method
     DEFAULT_FIELDS = [:id, :filename].freeze
 
     # Returns a list of Feedback Files associated with a group's assignment submission
-    # Requires: assignment_id, group_id
-    # Optional: filter, fields
+    # Requires: assignment_id, group_id OR test_run_id
+    # Optional: filter, fields, test_run_id
     def index
-      submission = Submission.get_submission_by_group_id_and_assignment_id(
-        params[:group_id], params[:assignment_id])
-
-      collection = submission.feedback_files
+      unless check_params
+        # incomplete/invalid HTTP params
+        render 'shared/http_status', locals: {code: '422', message:
+          HttpStatusHelper::ERROR_CODE['message']['422']}, status: 422
+        return
+      end
+      association = get_association_by_param
+      collection = association.feedback_files
 
       feedback_files = get_collection(collection) || return
 
@@ -24,18 +30,23 @@ module Api
         format.json { render json: feedback_files.to_json(only: DEFAULT_FIELDS) }
       end
       rescue ActiveRecord::RecordNotFound => e
-        # Could not find submission
+        # Could not find submission or test_run
         render 'shared/http_status', locals: {code: '404', message:
           e}, status: 404
     end
 
     # Sends the contents of the specified Feedback File
-    # Requires: assignment_id, group_id, id
+    # Requires: (assignment_id, group_id OR test_run_id), id
     def show
-      submission = Submission.get_submission_by_group_id_and_assignment_id(
-        params[:group_id], params[:assignment_id])
+      unless check_params([:id])
+        # incomplete/invalid HTTP params
+        render 'shared/http_status', locals: {code: '422', message:
+          HttpStatusHelper::ERROR_CODE['message']['422']}, status: 422
+        return
+      end
+      association = get_association_by_param
 
-      feedback_file = submission.feedback_files.find(params[:id])
+      feedback_file = association.feedback_files.find(params[:id])
 
       # Everything went fine; send file_content
       send_data feedback_file.file_content,
@@ -43,31 +54,28 @@ module Api
                 filename: feedback_file.filename,
                 disposition: 'inline'
       rescue ActiveRecord::RecordNotFound => e
-        # Could not find submission or feedback file
+        # Could not find submission, test run, or feedback file
         render 'shared/http_status', locals: {code: '404', message:
           e}, status: 404
     end
 
     # Creates a new feedback file for a group's latest assignment submission
     # Requires:
-    #  - assignment_id
-    #  - group_id
+    #  - assignment_id, group_id OR test_run_id
     #  - filename: Name of the file to be uploaded
     #  - mime_type: Mime type of feedback file
     #  - file_content: Contents of the feedback file to be uploaded
     def create
-      if has_missing_params?([:filename, :mime_type, :file_content])
+      unless check_params([:filename, :mime_type, :file_content])
         # incomplete/invalid HTTP params
         render 'shared/http_status', locals: {code: '422', message:
           HttpStatusHelper::ERROR_CODE['message']['422']}, status: 422
         return
       end
-
-      submission = Submission.get_submission_by_group_id_and_assignment_id(
-        params[:group_id], params[:assignment_id])
+      association = get_association_by_param
 
       # Render error if there's an existing feedback file with that filename
-      feedback_file = submission.feedback_files.find_by_filename(params[:filename])
+      feedback_file = association.feedback_files.find_by_filename(params[:filename])
       unless feedback_file.nil?
         render 'shared/http_status', locals: {code: '409', message:
           'A Feedback File with that filename already exists'}, status: 409
@@ -80,7 +88,7 @@ module Api
       else
         content = params[:file_content]
       end
-      if submission.feedback_files.create(filename: params[:filename],
+      if association.feedback_files.create(filename: params[:filename],
                                           mime_type: params[:mime_type],
                                           file_content: content)
                                           .valid?
@@ -95,12 +103,17 @@ module Api
     end
 
     # Deletes a Feedback File instance
-    # Requires: assignment_id, group_id, id
+    # Requires: (assignment_id, group_id OR test_run_id), id
     def destroy
-      submission = Submission.get_submission_by_group_id_and_assignment_id(
-        params[:group_id], params[:assignment_id])
+      unless check_params([:id])
+        # incomplete/invalid HTTP params
+        render 'shared/http_status', locals: {code: '422', message:
+          HttpStatusHelper::ERROR_CODE['message']['422']}, status: 422
+        return
+      end
+      association = get_association_by_param
 
-      feedback_file = submission.feedback_files.find(params[:id])
+      feedback_file = association.feedback_files.find(params[:id])
 
       if feedback_file.destroy
         # Successfully deleted the Feedback file; render success
@@ -118,19 +131,23 @@ module Api
     end
 
     # Updates a Feedback File instance
-    # Requires: assignment_id, group_id, id
+    # Requires: assignment_id, group_id OR test_run_id
     # Optional:
     #  - filename: New name for the file
     #  - file_content: New contents of the feedback file file
     def update
-      submission = Submission.get_submission_by_group_id_and_assignment_id(
-        params[:group_id], params[:assignment_id])
-
-      feedback_file = submission.feedback_files.find(params[:id])
+      unless check_params
+        # incomplete/invalid HTTP params
+        render 'shared/http_status', locals: {code: '422', message:
+          HttpStatusHelper::ERROR_CODE['message']['422']}, status: 422
+        return
+      end
+      association = get_association_by_param
+      feedback_file = association.feedback_files.find(params[:id])
 
       # Render error if the filename is used by another
       # Feedback File for that submission
-      existing_file = submission.feedback_files.find_by_filename(params[:filename])
+      existing_file = association.feedback_files.find_by_filename(params[:filename])
       if !existing_file.nil? && existing_file.id != params[:id].to_i
         render 'shared/http_status', locals: {code: '409', message:
           'A Feedback File with that filename already exists'}, status: 409
@@ -159,6 +176,27 @@ module Api
         # Could not find submission or feedback file
         render 'shared/http_status', locals: {code: '404', message:
           e}, status: 404
+    end
+
+    def check_params(required_params=[])
+      !(has_missing_params?(required_params) ||
+        (has_missing_params?([:assignment_id, :group_id]) &&
+          has_missing_params?([:test_run_id])))
+    end
+
+    def get_association_by_param
+      if params[:test_run_id]
+        association = TestRun.find(params[:test_run_id])
+      else
+        association = Submission.get_submission_by_group_id_and_assignment_id(
+          params[:group_id], params[:assignment_id])
+      end
+      association
+    end
+
+    def get_feedback_file
+      feedback_file = FeedbackFile.find(params[:id])
+      send_data_download feedback_file.file_content, filename: feedback_file.filename
     end
 
   end # end FeedbackFilesController
